+++ geotrek/tourism/urls.py
@@ -1,7 +1,7 @@
 from django.conf import settings
 from django.conf.urls import patterns, url
 
-from mapentity import registry
+from mapentity.registry import registry
 
 from geotrek.common.urls import PublishableEntityOptions